robot_body/robot_eyes/requirements.txt
flask>=2.3.0
flask-sock>=0.7.0
opencv-python>=4.8.0
numpy>=1.24.0
picamera2>=0.3.12
python-prctl>=1.8.1
websockets>=12.0
pyyaml>=6.0.1 